server/requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.12
pydantic==2.11.1
python-dotenv==1.1.0
Requests==2.32.3
uvicorn